Adding support for external examples.
The idea is that writing <example> in the actual code is very clumsy and messy.
We'd rather have external files that hold the examples, and can even be run separately, without using any ngdocs compiler.

We incorporate those external examples using a hint in the documentation itself:

@external-example:my-example

This will look for a directory named 'my-example' under the base examples directory (which can be configured in the Gruntfile).
Every file under that directory will become a <file> in the documentation.
The AngularJS module will have the name of the example, i.e. 'my-example'.

HTML files that are used this way can be real, full HTMLs.
The plugin will know to only take the <body> part and include it in the <file> section.

And - I had to get rid of the dependency on angular-animate, as this created bootstrapping problems, which I did not manage to understand or solve.
The very same HTML and JS files, when used as is externally, did not cause these problems :/